feat(api): refactor out path param, use oauth
